Plumb through channels
