+++ shared/chat/conversation/messages/index.js
@@ -24,7 +24,8 @@ const factory = (
   ) => void,
   onShowEditor: (message: Types.ServerMessage, event: SyntheticEvent<>) => void,
   isSelected: boolean,
-  measure: () => void
+  measure: () => void,
+  channelNames: {[string]: string}
 ) => {
   const kind = Constants.messageKeyKind(messageKey)
   switch (kind) {
@@ -41,6 +42,7 @@ const factory = (
       return (
         <Wrapper
           innerClass={Attachment}
+          channelNames={channelNames}
           isSelected={isSelected}
           measure={measure}
           messageKey={messageKey}
@@ -58,6 +60,7 @@ const factory = (
       return (
         <Wrapper
           innerClass={TextMessage}
+          channelNames={channelNames}
           isSelected={isSelected}
           measure={measure}
           messageKey={messageKey}
