+++ shared/chat/conversation/messages/index.js.flow
@@ -12,6 +12,7 @@ declare function factory(
   ) => void,
   onShowEditor: (message: Types.ServerMessage, event: SyntheticEvent<>) => void,
   isSelected: boolean,
-  measure: () => void
+  measure: () => void,
+  channelNames: {[string]: string}
 ): React.Node
 export default factory
