+++ shared/chat/conversation/messages/text/container.js
@@ -1,10 +1,5 @@
 // @flow
 import * as Constants from '../../../../constants/chat'
-import {
-  getConvIdsFromTeamName,
-  getTeamNameFromConvID,
-  getChannelNameFromConvID,
-} from '../../../../constants/teams'
 import * as Types from '../../../../constants/types/chat'
 import TextMessage, {type Props} from '.'
 import createCachedSelector from 're-reselect'
@@ -12,24 +7,8 @@ import {Set} from 'immutable'
 import {compose, lifecycle, connect, type TypedState} from '../../../../util/container'
 import {type OwnProps} from './container'
 
-const getTeamChannelNames = (state: TypedState, messageKey: Types.MessageKey): {[string]: string} => {
-  const message: ?Types.Message = Constants.getMessageFromMessageKey(state, messageKey)
-  if (!message || message.type !== 'Text') return {}
-
-  const teamname = getTeamNameFromConvID(state, message.conversationIDKey)
-  if (!teamname) return {}
-
-  const convIDs = getConvIdsFromTeamName(state, teamname)
-  return convIDs
-    .map(convID => {
-      getChannelNameFromConvID(state, convID)
-    })
-    .filter(Boolean)
-    .toObject()
-}
-
 const getProps = createCachedSelector(
-  [Constants.getMessageFromMessageKey, Constants.getEditingMessage, getTeamChannelNames],
+  [Constants.getMessageFromMessageKey, Constants.getEditingMessage],
   (message: ?Types.TextMessage, editingMessage, channelNames: {[string]: string}) => {
     return {
       isEditing: message === editingMessage,
@@ -37,7 +16,6 @@ const getProps = createCachedSelector(
       type: message ? message.messageState : null,
       mentions: message ? message.mentions : Set(),
       channelMention: message ? message.channelMention : 'None',
-      channelNames,
     }
   }
 )((state, messageKey) => messageKey)
