+++ shared/chat/conversation/messages/wrapper/container.js.flow
@@ -13,6 +13,7 @@ export type OwnProps = {|
   onShowEditor: (message: Types.ServerMessage, event: SyntheticEvent<>) => void,
   isSelected: boolean,
   innerClass: any,
+  channelNames: {[string]: string},
   measure: ?() => void,
 |}
